# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.Agent:
    properties:
      capabilities:
        items:
          type: integer
        type: array
      created_at:
        type: string
      endpoint:
        type: string
      failureCount:
        type: integer
      id:
        type: string
      lastError:
        type: string
      lastHeartbeatAt:
        type: string
      lastSyncFinishedAt:
        type: string
      lastSyncStartedAt:
        type: string
      name:
        type: string
      status:
        type: string
      tlsInsecure:
        type: boolean
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  router.agentListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.Agent'
        type: array
      total:
        type: integer
    type: object
  router.agentProbeResponse:
    properties:
      host: {}
      status:
        type: string
    type: object
...
  router.alertListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.Alert'
        type: array
      total:
        type: integer
    type: object
...
  router.createAgentRequest:
    properties:
      endpoint:
        type: string
      name:
        type: string
      tlsInsecure:
        type: boolean
      token:
        type: string
    type: object
...
  router.errorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: KVM 虚拟化资源管理控制台后端 API，提供认证、Agent 管理、运行态资源、刷新事件、任务、审计和告警接口。
  title: KVM Manager API
  version: "1.0"
paths:
  /api/agents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.agentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取 Agent 列表
      tags:
      - agents
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent 登记信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.createAgentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Agent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 登记 Agent
      tags:
      - agents
...
  /api/agents/test-connection:
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent 连接信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.createAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.agentProbeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 测试未登记 Agent 连接
      tags:
      - agents
  /api/alerts:
    get:
      parameters:
      - description: 告警状态，如 active、resolved
        in: query
        name: status
        type: string
      - description: 搜索关键词，可匹配级别、状态、标题、消息、来源和外部通知状态
        in: query
        name: q
        type: string
      - description: 告警元数据顶层字段名，配合 metadataValue 精确限定 JSON 字段；仅填写字段名时匹配存在该字段的告警
        in: query
        name: metadataKey
        type: string
      - description: 告警元数据字段值关键词；metadataKey 为空时在整段元数据 JSON 中模糊搜索
        in: query
        name: metadataValue
        type: string
      - description: 返回数量，可选 30、50、100、200 或 all
        in: query
        name: limit
        type: string
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.alertListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警列表
      tags:
      - alerts
...
  /api/audit-logs:
    get:
      parameters:
      - description: 搜索关键词，可匹配动作、用户、资源、IP 和元数据
        in: query
        name: q
        type: string
      - description: 审计元数据顶层字段名，配合 metadataValue 精确限定 JSON 字段；仅填写字段名时匹配存在该字段的日志
        in: query
        name: metadataKey
        type: string
      - description: 审计元数据字段值关键词；metadataKey 为空时在整段元数据 JSON 中模糊搜索
        in: query
        name: metadataValue
        type: string
      - description: 返回数量，可选 30、50、100、200 或 all
        in: query
        name: limit
        type: string
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.auditLogListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取审计日志列表
      tags:
      - logs
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 使用本地账号或已启用的外部认证方式（如 AD/LDAP）登录，返回会话 Token 和用户信息。企业微信直连与统一认证中心不走密码登录，需使用对应的
        OAuth 发起接口。默认空库管理员为 admin / 123456。
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: 登录
      tags:
      - auth
  /api/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 注销当前会话
      tags:
      - auth
  /api/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.meResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户
      tags:
      - auth
  /api/auth/password:
    put:
      consumes:
      - application/json
      description: 校验旧密码后更新当前登录用户密码。新密码至少 6 个字符，且不能与旧密码相同。
      parameters:
      - description: 密码信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 修改当前用户密码
      tags:
      - auth
  /api/auth/password-reset/captcha:
    get:
      description: 返回登录页找回密码流程使用的算式验证码与签名 Token。图形验证码有效期由基础配置控制。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.passwordResetCaptchaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: 获取找回密码图形验证码
      tags:
      - auth
  /api/auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      description: 校验找回密码验证码后重置本地账号密码，并清理该用户已有会话。
      parameters:
      - description: 验证码与新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.passwordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: 确认找回密码并重置密码
      tags:
      - auth
  /api/auth/password-reset/send-code:
    post:
      consumes:
      - application/json
      description: 使用已启用找回密码用途的邮件媒介发送找回密码验证码。发送前必须携带用户名与图形验证码校验得到的短期 Token，并校验验证邮箱与账号配置邮箱一致；验证码会发送到账号配置邮箱，不使用邮件通知配置中的告警收件人列表。验证码有效期、发送冷却和频率限制统计窗口由基础配置控制。
      parameters:
      - description: 发送配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.passwordResetSendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.passwordResetSendCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/router.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: 发送找回密码验证码
      tags:
      - auth
...
  /api/auth/providers:
    get:
      description: 返回已启用的外部认证方式；本地账号登录始终可用。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.publicAuthProviderListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: 获取登录页可用认证方式
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"